requirements.txt
PyYAML
# chinese
jieba
pypinyin
ToJyutping
opencc-python-reimplemented 